+++ app/renderer/pages/devPage/index.js
@@ -46,8 +46,10 @@ const DevPage = ({
   onGetSessionProgress,
   onGetEraLength,
   onGetIntentionsBalances,
+  onToggleGlobalModal,
   onStakeAndRestart,
   onUnStakeAndRestart,
+  wallets,
 }) => {
   const {
     localNode: { chain },
@@ -67,15 +69,20 @@ const DevPage = ({
         </PageTitle>
         <Flex>
           <Button onClick={() => onTestToaster()}>Toaster Test and Reset App UI state</Button>
+          <Button
+            onClick={() =>
+              onToggleGlobalModal({ isOpen: true, type: 'EXIT_APP_WHILE_STAKING_MODAL' })
+            }
+          >
+            Global Modal
+          </Button>
         </Flex>
         <Flex>
           <Button
             onClick={() =>
               onStake({
-                stashAccountAddress: Object.keys(
-                  window.odin.store.getState().localStorage.WALLETS[0].accounts
-                )[0],
-                wallet: window.odin.store.getState().localStorage.WALLETS[0],
+                stashAccountAddress: Object.keys(wallets[0].accounts)[0],
+                wallet: wallets[0],
               })
             }
           >
@@ -84,10 +91,8 @@ const DevPage = ({
           <Button
             onClick={() =>
               onUnStake({
-                fromAddress: Object.keys(
-                  window.odin.store.getState().localStorage.WALLETS[0].accounts
-                )[0],
-                wallet: window.odin.store.getState().localStorage.WALLETS[0],
+                fromAddress: Object.keys(wallets[0].accounts)[0],
+                wallet: wallets[0],
               })
             }
           >
@@ -98,10 +103,8 @@ const DevPage = ({
           <Button
             onClick={() =>
               onStakeAndRestart({
-                wallet: window.odin.store.getState().localStorage.WALLETS[0],
-                stashAccountAddress: Object.keys(
-                  window.odin.store.getState().localStorage.WALLETS[0].accounts
-                )[0],
+                wallet: wallets[0],
+                stashAccountAddress: Object.keys(wallets[0].accounts)[0],
               })
             }
           >
@@ -147,7 +150,7 @@ const DevPage = ({
         <Flex>
           <Button
             onClick={() =>
-              window.odin.api.cennz.createWalletWithHDKeyRing({
+              window.appApi.createWalletWithHDKeyRing({
                 passphrase: '',
               })
             }
@@ -158,7 +161,7 @@ const DevPage = ({
         <Flex>
           <Button
             onClick={() =>
-              window.odin.api.cennz.restoreWallet({
+              window.appApi.restoreWallet({
                 mnemonic:
                   'dove pull aerobic reason husband electric egg ceiling castle swear tank proud',
                 passphrase: '',
@@ -176,7 +179,7 @@ const DevPage = ({
                 fromAddress: '5Gw3s7q4QLkSWwknsiPtjujPv3XM4Trxi5d4PgKMMk3gfGTE',
                 toAddress: '5F1XzAhQGNcapqm666QNU2cMUnj9DRAKB6vGXoVTkxWMLhPs',
                 amount: new BN('13', 10),
-                wallet: window.odin.store.getState().localStorage.WALLETS[0].wallet,
+                wallet: wallets[0].wallet,
               })
             }
           >